+++ datamuru/cli/commands/__init__.py
@@ -5,7 +5,6 @@ from .edition import edition_group
 from .init import init_command
 from .import_ import import_group
 from .plan import plan_command
-from .ui import ui_command
 from .validate import validate_command
 
 COMMANDS = [
@@ -17,5 +16,4 @@ COMMANDS = [
     doctor_command,
     import_group,
     edition_group,
-    ui_command,
 ]